Redraw GtkHex widget when the HexDocument changes (Fix for bug #500358)
Connect to the HexDocument's document_changed signal and redraw the widget so
that the display gets updated when modifying the data with
hex_document_set_data().
Since the GtkHex view is connecting to the 'document_changed' signal, we no
longer have to rely on the HexDocument object to emit the 'data_changed' signal
for each view in the gh->views list.

svn path=/trunk/; revision=1177