fix a bug where the selection would always start at 0 if
* src/gtkhex.c: fix a bug where the selection would always start at 0 if

the user only used the keyboard.  To reproduce, navigate away from the start
of the document, then press Shift+Down, and the text will be selected, but
instead of starting the selection from where the cursor was when you pressed
Down, it would start from the beginning of the file.

svn path=/trunk/; revision=1206