Fix regression with close on save confirmation
When closing an open file, one of the checks that is done is to see if
the file still exists on the disk. However, d48fdd1 introduced a
regression so that this check always returns true, making ghex always
offer to save the file.

Fixed that check to use g_file_new_for_path(), as we are dealing with
local pathnames. While at this, also renamed the function to avoid
future confusion and made it static.