ghex.desktop: Use 'Hex Editor' as the generic name
It is a more widely used term than 'Hexadecimal Editor' (see e.g.
http://en.wikipedia.org/wiki/Hex_editor) and openSUSE is already
patching the desktop file to use this; it is also consistent with last
ghex stable release that had 'Hex Editor' for the Name tag.

As a plus, we currently have twice as many translations for 'Hex
Editor', compared to 'Hexadecimal Editor'.

While at this, also improved wording for the Comment tag.