Use g_uri_list_extract_uris when a file is dragged and dropped
When files are dragged and dropped into GHex, the drag data comes
through as text/uri-list. This media type is described by RFC 2483 and
can actually be quite complex, with CRLF at the end of every line and
optional comment lines interspersed with the data. Without using the
right function to parse this data, GHex shows an error message when
dropping files from PCManFM to GHex.

https://bugzilla.gnome.org/show_bug.cgi?id=679717