Distribute extra empty space between offsets and xdisp
If offsets are shown, distribute the extra empty space among the
gap between the offsets and xdisp and the gap between xdisp and
adisp instead of allocating it all to the second gap.

Allows better visual separation of offsets from the first xdisp
column on themes that do not have any padding (e.g., adwaita on
Fedora 28).