Implement tabbed viewing and file opening.
Closing tabs not yet implemented, nor is saving, so beware.

I moved the scrollbar back to the GtkHex widget. It is not appearing,
likely due to bugs of overdrawing I need to work on when I switch gears
back to the widget. Scrolling still works, but the scrollbar does not
visibly appear.