gtkhex: fix cut, copy and paste behaviour
This is a backport of a fix from the ghex4 WIP.

The range for copy/pasting was being miscalculated and was always off by
one character, resulting in one character being missing from copy/paste
operations.

As well, it was not previously possible to cut or copy a single
character at a time. This has been implemented in this commit as well.