Rewrite render_cursor
This should get the cursor rendering to work properly on both X11 and
wayland now. The problem was, before gtkhex was basically guessing where
to draw the cursor based on some janky assumptions, and then re-drawing
the hex or ascii characters atop the drawing area.

This is, of course, not the best approach, so the rewrite just uses the
existing data from the pango layouts and draws the cursor using that
data.