gtkhex: Rewrite render_cursor
This should de-jankify highlighting on wayland and keep it at parity
with X11.

There were similar issues to the cursor rendering. Assumptions were made
based on multiplications of calculated character widths to decide the
x/y coordinates and widths of the highlights.

This, again, was a broken approach. We now just pull that information
from pango and gdk, and render as necessary.