gtkhex: Make context menu a child of the gh widget
This corrects some bizarro size allocation issues that resulted in a
scrollbar appearing unnecessarily on the popover. Previously, a new
context menu widget was built each time a right-click occurred, and was
made a child of the respective hex/ascii drawing area as opposed to the
gh widget as a whole.