findreplace: Add clipboard keybindings.
This patch adds Ctrl+{C,X,V} keybindings to the Find, Replace and
Advanced Find dialogs of ghex3.

Note that this has been done manually in these dialogs as opposed to
associating them with the widget per (lib)g(tk)hex4. That is simply
because some other programs actually use the libgtkhex3 widget
separately, and I do not want to modify its expected behaviour at this
time as it is legacy software.

Fixes #36