gtkhex: Cursor rendering improvements
No longer simply render based on :selected bg color - use :checked
pseudoclass (if we'd used separate widgets for all this stuff we
wouldn't have to be hunting for pseudoclasses that _kinda_ fit...)

Draw a dashed cursor for either pane if out of focus.

ISSUES: some redraws getting missed when widgets go in and out of focus.
Likely this isn't a _new_ issue, but just having these different colours
is revealing issues that were there all along.