gtkhex: Bugfixes to layout manager.
This allows us to lay out the widget with CSS and not worry about any
characters being cut off.

Next step in this avenue would be to tweak further, especially the
allocation - not of the widget itself, but of the number of characters
that can be drawn on the hex widget.