appwin: Turn on insert mode on new file action
This is subject to change on a whim as it could be annoying.

But right now after some "feel" testing, it just feels wrong to me to
click File->New and not be able to type right away. Although, some
people (myself included) use a particular editor that behaves this way
all the time...