hex-doc rewrite with iface: INITIAL COMMIT.
This is stage one in my efforts to implement a memory-mapping backend
for ghex.

In this commit, we streamline HexDocument considerably and add a new
interface called HexBuffer. The legacy `malloc` gap buffer system has
been implemented as HexBufferMalloc. This interface can be swapped out
for another without breaking ABI in gtkhex, so this will allow for a new
backend to be written while keeping this one as well -- in future, it
could be used for backwards compatibility, or for smaller files.

I have started another backend but it is not yet working code, so it is
committed so it follows along, but is disabled and likely won't compile
at this time, let alone do anything.

I'm sure some further cleanups will happen as well as some further
bugfixes as I notice them. Since this required some substantial
refactoring of the backend, some bugs I have not caught may have slipped
through.

Please test!