API: Standardize datatypes and signedness
Standardize on gint64 so that we have a minimum cursor pos and file size
that works with files larger than 2 GB.

This doesn't fix all issues. nb: Find/Replace needs to be optimized. It
gets OOM-killed when searching for a string towards the end of the
Fedora Core 35 ISO, for instance.