Optimizations for large files and mmap tweaks
There were still many examples of 32-bit datatypes being used when
larger datatypes were needed. Review code and ensure consistency. It is
possible there are still a few bugs but it seems to be in working order
now.

For gtk_hex_update_all_autohighlights, limit the rendering to the
previous and next 10 screens. This speeds things up *immensely* (read:
makes it actually usable) with large files. This `10` number is
arbitrary and can be changed if it feels unnaturally low.

Unlink tmpfile name pre-emptively (mmap)