HexWidget: fix offsets redraw issue
Inaccurate number of lines of offsets would be drawn when the document
was first loaded.

Another thing was that any callbacks on state flags changed has been
removed. This should probably not have been there to begin with, but did
force some redraws. Removing that could cause other redraw issues I'm
not aware of. So will need to monitor this and make sure things work as
expected.