help: add support for static HTML help
This is a fallback for systems that do not have Yelp installed.

Use `-Dstatic-html-help=true` to enable.

<rant>
Other projects, take heed! This took a lot of code/effort to make help
displayable on both KDE, GNOME and other DE's/WM's. It's really
unnecessary that it's this hard to get Help to display on both KDE and
GNOME, given that *both* systems can display HTML help, and *both*
systems can display docbook.

It's super silly that both KDE and GNOME use help:/// as their URI
prefix schemes, but incompatibly.
</rant>