gtkhex: Segregate mmap backend as plugin, etc.
I didn't want to make this sweeping of a change between beta and stable,
but I think the work needed to be done before the initial stable release
of the library was complete.

The mmap buffer being a compile-time option only for 4.0 is fine, but we
can't have divergent APIs due to compile-time differences.

Add new hex_buffer_util_new function to spin up a valid implementation
of a HexBuffer interface, to abstract the API from the underlying
implementations. The `malloc` buffer is now baked into the library as a
fallback.

Make the requisite changes to the meson scripts.

The versioning has also been pinned down in this commit. The initial
stable release of GHex the application will be 42.0 and will include
libgtkhex 4.0.0.