widget: Support PRIMARY selection
Un-highlight when another PRIMARY selection is made

Support middle-click paste for PRIMARY

Get rid of the distinction between :selected and :focus-within
colouring in the CSS stylesheet -- now that PRIMARY is used to de-select
selections in a similar fashion to GtkText & friends, we don't see as
many things appearing constantly highlighted as we switch from one hex
widget to another.

This may be revisited in future, as I do like the
light grey appearance when focus is away from the highlights; but the
kink I can't work out right now is that the context-menu activating
*also* makes the highlighting show up in the grey colour, which looks
almost invisible on dark mode, which is a usability issue I want to
avoid more-so than I want the unfocused selections to appear grey.