hex-dialog: Prevent decimal display widgets from getting truncated
We used to calculate the number of allowed characters in the decimal
widgets by converting a large number to a string representing the
decimal value and counting the number of characters in that string.
Previously we used 0xfbfbfbfbfbfbfbfb as the "large number" for all
types of ints: signed/unsigned 8-bit all the way to signed/unsigned
64-bit.

This number is not long enough for some types of integers. For example,
for a signed 8-bit number, 0xfb is -5, which is only 2 characters. This
causes all uint8 numbers to get truncated to the first 2 characters
(-128 will get truncated to -1).

This is also not appropriate if the user enables "Show unsigned and
float as hexadecimal", since the "0x" in front of hexadecimal numbers
makes them even longer than decimal numbers for small bit widths.

The widget character limit has been removed. The reason to have them no
longer applies. According to 441f37f9:

> hex-dialog: Recalculate entry sizes depending on chosen conversion
>
> This (a) makes the type conversion dielog more compact and easier to
> read, and (b) decreases the minimum width of the whole GHex window
> enough so that we can display the hex view in 16 columns again.

Recently (at least since GTK 4), this is not true. The character limit
of these widgets does not affect the width of the widgets or the minimum
width of the window.