Fix erroneous error message when writing empty files
The user might try to save an empty file for at least two reasons:

1. The user opened an empty file, made some changes, undid them, then
   saved the same empty file.
2. The user opened a non-empty file, deleted all contents, then saved
   the now empty file.

Previously when trying to save an empty file in either of these cases,
GHex would show an error dialog that said, "There was an error saving
the file."

In reality, the file was saved/truncated successfully. GHex failed to
detect this situation as successful.

In malloc the malloc-based file buffer, GHex uses a gap buffer. The
write is broken into steps:

1. fopen the file in write mode, truncating it.
2. Write the buffer up to the gap, if one exists.
3. If there is a gap, move past the gap and write the second portion of
   the file.

Previously, the success of the write operation was whether steps 2 and 3
were both sucessful (or, if there was no gap, whether step 2 alone was
successful). However, for a zero-length buffer, we skip steps 2 and 3,
so the success of the operation was left at the default FALSE. In fact,
step 1 is also part of the write operations, and for a zero-length
buffer, opening a file to truncate it and then closing it is all that is
required.

Now we take the success of step 1 into consideration when deciding
whether the write was successful. In other words, the write was
successful if:

	(step 1 succeeds) && (step 2 succeeds) && (step 3 succeeds)