Enable/disable save and revert actions based on external file changes
If the open file changes due to some other process writing to it, the
"Revert" and "Save" buttons will get enabled if they aren't already. I didn't
make any direct changes to the UI, but I left open the future possibility of
showing a message like "The file has changed, would you like to reload it?"

This also fixes a weird thing about the current behavior: reverting a
file previously did not reset the "document has been changed" flag. Now
it does.

I also noticed that making a change and then undoing it does not clear
the flag. I didn't change this behavior.

Part of #67