A feature for pasting hex strings with or without delimiters. Also, the delimiter can be multiple spaces, multiple tabs or multiple new-line characters.
At present, GHex only allows pasting hex strings where bytes are
delimited by spaces. This change brings a way to
paste hex strings with or without a delimiter for bytes. The delimiter
can be multiple spaces, multiple tabs or multiple new-line
characters. A pasted string can contain bytes that are delimited
and not delimited, and can also use different delimiters.

One other change is not using the `sscanf` function for getting
the byte value from a hex string. The reason is that `sscanf` stops
reading its input after it encounters an invalid hex digit.
As a consequence of this, if you try to paste `FA AC 23 3K DD 91`,
what happens is the string `FA AC 23 03 DD 91` gets pasted.
Since K is not a valid hex digit, it only reads up to the digit `3`
while reading the byte `3K`. Now, if you try to paste something like this,
an appropriate error will be set, which later results in showing an error
message to the user.

Fixes #89