ci: Do not explicitly define tarball-artifact-path
If left to default it will now discover the tarball automatically
as long we export the public-dist directory as the job artifact.
(which the flatpak template does already)

Note that this change which was previously optional may now be
necessary. Thanks to alateria.

(cherry picked from commit 1fb401945652fed5289a9f5b9a34214ea5ac710b)